+++ NEWS.md
@@ -1,5 +1,12 @@
 # Change log
 
+## [zen4R 0.10.3](https://github.com/eblondel/zen4R) |
+[![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-unavailable-red.svg)](https://cran.r-project.org/package=zen4R)
+
+**New features**
+
+* [#179](https://github.com/eblondel/zen4R/issues/179) Support method to get a record citation
+
 ## [zen4R 0.10.2](https://github.com/eblondel/zen4R) |
 [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/package=zen4R)